Name Change and 0.7.3.8 Update
0.7.3.0: Changed name to Exodus
0.7.3.1: Fixed a bug with the rebuild system that made broken ISOs.
(11/22)
0.7.3.2: Added ModHandler class, setting up for addition of the modding
interface (11/25)
0.7.3.3: Added UI for ModHandler class. "Apply" button currently does
nothing (11/26)
0.7.3.4: Created a separation between Exodus data classes and
Transformers game data classes. More classes need to be added to this
system but the separation exists. This is the first step for Randomizer
logic changes in future updates. (11/30)
0.7.3.5: Changed the rebuild system so that it no longer needs to copy
files before zipping them. STILL NEEDS TESTING and to be copied over for
the SOUNDE zip (12/1)
0.7.3.6: Testing for previous change confirmed. SOUNDE zip added to
process. Similar system is being worked on for the ISO portion of the
rebuild, but needs testing. (12/2) (12/2 evening: multiple tests confirm
- with how imgburn handles paths, folders, and file overwrites, this
doesn't work for files outside the zips. need to find another approach.)
0.7.3.-: (there was a fix for the imgburn issue between these. Changelog
was not updated due to unintended file deletion while working on the
fix.)
0.7.3.7: ModHandler class now displays which folders will be pulled from
the unmodified source and which will be rebuilt from the modded files.
This integrates with file replacements to keep track of which folders
need to be taken from the modified files. (12/8)
0.7.3.8: Additional changes to ModHandler and ISOBuilder to accommodate
the ModHandler system. The "Apply Changes" button on the Mods and
Replacements interface now builds an ISO with the settings selected by
the user. (12/9)