0.7.3.9 Update - Further Class Separation
More work into separating the Transformers and Exodus data classes.
There's still quite a bit  to do, including:

The constructor for taPickupPlaced still needs to be defined.
Data that used to be part of the Minicon class (team, powercost) is
still referenced by functions. Some way to connect the Exodus data
classes with their TFA counterparts needs to be created.