Addition of Boolean to taData Value Function
The taData value() function can now return boolean types. While
attempting to add std::vector<int> to this function for the IntegerArray
type, I found that using the varient and std::get system will not work
easily for this. Significant changes will be needed for this system to
return all types.