More Database Editing Improvements
Feature addition: When submitting an edit, the Instance that was edited
will be selected when the TreeView reloads

Bugfix: Corrected a memory leak caused by the TreeView not being tracked
by ProgWindow, creating a new object every time an edit was made or
another file was selected. This would quickly add up.