Multiple Changes for Database System
1. Corrected the initialization of the maxInstances value to be the
highest value of an instance instead of the count. This will prevent
duplicate instance indecies when adding to a file.

2. Matching change made to the loadRequiredFile function.

3. Changed the copyOrDeleteInstance function to directly use the
QModelIndex instead of the instance index. This prevents the wrong
instance from being copied or deleted.

4. Commented out some debugging lines to speed up file loading while
debugging. These sections have been thoroughly tested now.

5. Added signal to set the dataTree pointer to nullptr when the datatree
is deleted. This was done to prevent crashes caused by the
selectionModel check still being run if the user loaded this file, did
not make a selection, then loaded a different view and returned to this
file. However, this seems to have introduced an inconsistent crash when
closing the program. This seems harmless for now, so I'm not pursuing a
better solution yet. I think this may be caused by the order in which
things get cleaned up as the program closes.