Bugfix: Crash Caused by Endgame
Cut Minicons, including Endgame, will be present in the Exodus data but
are not consistent in the METAGAME file. The testing file had Endgame's
ID as 31, another as 13, and another as 0. To account for this mismatch,
this logging loop will simply skip over it if a match is not found.

NOTE: This bugfix will need to be added in other areas that call the
getGameMinicon function.