The Great Refactoring (Part 1)
In an effort to incorporate the design principles that I've learned
since starting this project, I've gone back and changed the structure of
the project. The key points are as follows:

1. The ProgWindow class no longer exists. It has been divided into the
exWindow, exDebugger, and zlManager classes. Eventually, these will
follow a one class one file structure, but this is only part 1.

2. The Winsparkle updater has been given its own class.

3. The exWindow and exDebugger classes are single instance classes,
others are passed via dependency injection.

4. All class systems will be responsible for adding their own UI
elements and menu bars instead of the UI class needing to manage all of
them.

5. Program initialization now happens in main instead of in the UI
class's constructor.

At this time, a lot of surface-level functionality has been lost.
Changes 4 and 5 are still in progress. Once these are finalized, all
previously existing functions should be unchanged.