The Great Refactoring - Part 1.5
After further planning and reflection, I realized that the UI should not
be a singleton class. This will allow for easier implementation of
popout windows later on.

I think there are still some issues with the implementation of
exWindowBase and exWindowMain's constructors at this time, but they
appear to be functional. The small graphical issues at startup will be
addressed another time.