Changes to FileData Processing
Currently, FileData uses multiple "read<type>" functions to read data
from the buffer, while writing is left to BinChanger. This is changing
to a single "process" function with multiple overloads. Once this change
is made throughout the project, the responsibility for writing will
change to FileData, which will simply write the contents for dataBytes
to the file.