ITF Rewrite for FileData process Improvements
This contains a rewrite of the ITF system. The main data of the ITF type
is no longer contained in the QImage class, instead using a container of
the basic data. This is converted to a QImage when needed for display
purposes.

While this wasn't necessarily required to make the change to reading and
writing via a single function, I liked this solution more. It also
simplified some functions in ways I'm happy about.

Currently, the only function that hasn't been updated is
convertColorToIndex. This will be brought back in a later update.