ITF Complete Process Changeover
This finalizes the changes needed to make the ITF format work with the
new FileData process system. It also introduces some bugfixes and
corrects the data written to ITF files.

1. All read functions have been changed to process functions. These now
properly resize their containers and pass objects to process instead of
creating objects that get pushed into their containers.
2. The swizzle function has been given enum values to make it clearer
whether the data is being mixed or not
3. The swizzle function will no longer mix an already mixed image or
attempt to unmix an ordered image
4. Alpha values are only adapted when converting alpha types or in
images created for display. This preserves the correct value to write to
ITF files.
5. Corrected palette processing to use the correct data type
6. Moved 16bpp processing to FileData