Project Restructuring for Shared Files
The organization of the entire project has been rearranged and the build
type has been changed to cmake. This is to facilitate two things:

1. Shared files between multiple projects, specifically the UI,
debugging, and binary and file operations.

2. The addition of the YAML library I plan to use. I had to either move
the winsparkle incorporation from qmake to cmake or the YAML
incorporation from cmake to qmake, and I decided this would be the
better path.

This change will also be made to SELF. I'm currently trying to figure
out the best course of action to keep the common files tracked. For now,
they're excluded from this repository.